fix: 局部拦截器不支持的 bug

+++ grpc-starter/grpc-starter-spring-boot2-server-ram-redis/src/main/java/io/github/jojoti/grpcstartersbramredis/AbstractSessionUser.java
@@ -36,7 +36,7 @@ final class AbstractSessionUser implements SessionUser {
         this.validToken(token, attachInline);
     }
 
-    protected AbstractSessionUser(TokenDAO tokenDAO, ObjectMapper objectMapper, long uid, int scopeId, ImmutableList<String> attachInline) {
+    protected AbstractSessionUser(TokenDAO tokenDAO, ObjectMapper objectMapper, long uid, long scopeId, ImmutableList<String> attachInline) {
         this.tokenDAO = tokenDAO;
         this.objectMapper = objectMapper;
         final var session = tokenDAO.getSession(uid, scopeId, attachInline);
